requirements.txt
# Core dependencies
python-dotenv>=0.21.0
discord.py>=2.3.2
requests>=2.28.0
pymongo>=4.0.0
pathlib
Pillow>=10.0.0  # For image processing
aiohttp>=3.9.0  # For async HTTP requests
urllib3>=1.21.0  # For HTTP requests
tiktoken>=0.4.0
colorama>=0.4.6
# test
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0